Preparation for introducing Tile collision
Although it has introduced some stuttering of the yetis, when they move
on the diagonal.

The actual movement of the units, is now being performed by the object
handler, and no longer needs to be place inside of the Update() method.

Minor : fixed bug with Menu default coordinates of screen width and
height / 2 being set to fixed numbers.