Basic Makefile that should work on Linux
Based on http://www.puxan.com/web/blog/HowTo-Write-Generic-Makefiles
It should work fine on Linux systems, but probably not out of the box on Windows
since it presupposes g++ as C++ compiler, and uses a Unix shell function to list
all *.cpp files recursively. It can most likely be improved later on to also work
for other systems though, if someone has the knowledge to do it.